+++ src/api/axiosInstance.js
@@ -1,12 +1,8 @@
 import axios from 'axios';
 
-const BASE_DOMAIN = 'http://api.richardktran.dev/api';
+const BASE_DOMAIN = 'http://api.yourhometonight.com/api';
 const baseURL = BASE_DOMAIN;
 
 export const axiosInstance = axios.create({
   baseURL,
-  timeout: 1000,
-  headers: {
-    'Content-Type': 'application/x-www-form-urlencoded',
-  },
 });
