+++ src/api/loginApi.js
@@ -2,17 +2,10 @@ import { axiosInstance } from './axiosInstance';
 
 const loginApi = {
   login: (user) => {
-    axiosInstance
-      .post('/login', {
-        email: user.email,
-        password: user.password,
-      })
-      .then((response) => {
-        return response;
-      })
-      .catch((err) => {
-        console.log(err);
-      });
+    return axiosInstance.post('/login', {
+      email: user.email,
+      password: user.password,
+    });
   },
 };
