--- src/components/LargeButton/LargeButton.jsx
@@ -1,14 +1,0 @@
-import { Button } from 'antd';
-import React from 'react';
-
-import './LargeButton.scss';
-
-const LargeButton = () => {
-  return (
-    <Button className="button__wrapper" type="primary">
-      Click me
-    </Button>
-  );
-};
-
-export default LargeButton;
