+++ src/components/LoginForm/LoginForm.scss
@@ -1,0 +1,34 @@
+.login {
+  &__form {
+    width: 100%;
+    &__title {
+      display: block;
+      font-weight: bold;
+      font-size: 39px;
+      color: #333333;
+      line-height: 1.2;
+      text-align: center;
+    }
+
+    &__link {
+      display: flex;
+      justify-content: flex-end;
+      align-items: center;
+
+      span {
+        margin-right: 20px;
+      }
+
+      a {
+        font-size: 16px;
+      }
+    }
+
+    .submit__button {
+      width: 100%;
+      border-radius: 50px;
+      padding: 0 20px;
+      height: 50px;
+    }
+  }
+}
