+++ src/pages/HomePage/HomePage.jsx
@@ -1,15 +1,9 @@
 import React from 'react';
 
-import { LargeButton } from '../../components';
 import './HomePage.scss';
 
 const HomePage = () => {
-  return (
-    <div className="homepage__container">
-      HomePage
-      <LargeButton />
-    </div>
-  );
+  return <div className="homepage__container">HomePage</div>;
 };
 
 export default HomePage;
